SoC
remove unused

waveform_controls -> settings

add lic
